global variables on window
remove comments
